break out benchmarks into separate file create examples file for better godocs
